analyzer/visualizer: show 95.00% and 99.99% lat data
Currently we have 99.00% latency data only in reults summary page.
It might better to show 95.00% and 99.99% latency data also as these
data are also valuable for tail latency study.

This patch does the calculation the 95.00% and 99.99% data and alow
these data to be displayed on the summary page. Also we modified the
index page to display avg latency and 95.00% 99.00% and 99.99% latency data.

Signed-off-by: Yuan Zhou <yuan.zhou@intel.com>
Signed-off-by: Chendi.Xue <chendi.xue@intel.com>